rename trajectory analysis
