+++ python/src/worker/tasks/factory.py
@@ -16,7 +16,7 @@ from .list_genes import ListGenes
 from .marker_heatmap import MarkerHeatmap
 from .mitochondrial_content import GetMitochondrialContent
 from .pseudotime import GetPseudoTime
-from .trajectory_graph import GetTrajectoryGraph
+from .starting_nodes import GetStartingNodes
 
 
 class TaskFactory:
@@ -34,7 +34,7 @@ class TaskFactory:
             GetMitochondrialContent,
             MarkerHeatmap,
             GetPseudoTime,
-            GetTrajectoryGraph,
+            GetStartingNodes,
             GetExpressionCellSets,
         )
     }
