+++ r/NAMESPACE
@@ -15,7 +15,6 @@ export(getExpressionCellSet)
 export(getExpressionValues)
 export(getList)
 export(getMitochondrialContent)
-export(getPseudoTime)
 export(getTopMarkerGenes)
 export(load_experiment_data)
 export(makePseudobulkMatrix)
@@ -24,9 +23,10 @@ export(runDE)
 export(runDotPlot)
 export(runEmbedding)
 export(runExpression)
-export(runGenerateTrajectoryGraph)
 export(runMarkerHeatmap)
+export(runPseudoTimeTask)
 export(runPseudobulkDE)
+export(runStartingNodesTask)
 export(sendCellsetToApi)
 export(subsetIds)
 export(updateCellSetsThroughApi)
