+++ r/man/runPseudoTimeTask.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/trajectory_analysis.R
-\name{getPseudoTime}
-\alias{getPseudoTime}
+\name{runPseudoTimeTask}
+\alias{runPseudoTimeTask}
 \title{Calculate pseudotime}
 \usage{
-getPseudoTime(req, data)
+runPseudoTimeTask(req, data)
 }
 \arguments{
 \item{req}{{body: {
