+++ r/man/runStartingNodesTask.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/trajectory_analysis.R
-\name{runGenerateTrajectoryGraph}
-\alias{runGenerateTrajectoryGraph}
+\name{runStartingNodesTask}
+\alias{runStartingNodesTask}
 \title{Generate node coordinates for the initial trajectory analysis plot}
 \usage{
-runGenerateTrajectoryGraph(req, data)
+runStartingNodesTask(req, data)
 }
 \arguments{
 \item{req}{list of configuration parameters}
